Validate P3F ELF/SLUS by its header, not extension.
PS2 SLUS executables do not normally end in ".elf" and forcing a particular extension is not always a reliable way to ensure the file data is the correct format. Instead, let's allow all files to be used, and instead validate that the ELF header is present in the first four bytes of file data. This makes it easier to support building mods for real PS2 hardware. Because of issues with the project including non-existant PDB files for its build process (likely due to an oversight with your gitignore file ignoring PDBs), I am unable to test these changes.