Remove missing PDB files from project
Due to this repo's gitignore stripping any PDBs from being included in commits, it is erroneous to require they be included as part of the build, as that makes it impossible for people to clone the repo and build it without either manually building the dependencies and copying over the PDB files, or manually editing the project file to remove the project's dependence on them. I have opted to do the latter, as most of the other libraries this project depends on do not require PDB files to be copied over, and they shouldn't be necessary anyway as they are simply debug symbols, and removing them from the build will make subsequent releases smaller in size.